# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/products": {
      "post": {
        "operationId": "ProductsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Create product"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "products"
        ]
      },
      "get": {
        "operationId": "ProductsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List products"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "products"
        ]
      }
    },
...
    "/analytics/general": {
      "get": {
        "operationId": "AnalyticsController_getGeneralAnalytics",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get unified general analytics dashboard payload"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "analytics"
        ]
      }
    },
...
    "/analytics/products": {
      "get": {
        "operationId": "AnalyticsController_getProductsAnalytics",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get product analytics table"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "analytics"
        ]
      }
    },
    "/analytics/operations": {
      "get": {
        "operationId": "AnalyticsController_getOperationsAnalytics",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get operations analytics dashboard payload"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "analytics"
        ]
      }
    },
    "/analytics/payments/advanced": {
      "get": {
        "operationId": "AnalyticsController_getPaymentsAdvancedAnalytics",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get advanced payments analytics payload"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "analytics"
        ]
      }
    },
    "/analytics/financial": {
      "get": {
        "operationId": "AnalyticsController_getFinancialAnalytics",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get financial analytics dashboard payload"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "analytics"
        ]
      }
    },
...
    "/analytics/shipments": {
      "get": {
        "operationId": "AnalyticsController_getShipmentsAnalytics",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get legacy delivery analytics dashboard payload"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "analytics"
        ]
      }
    },
    "/analytics/reports/customers.csv": {
      "get": {
        "operationId": "AnalyticsController_exportCustomersReportCsv",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "analytics"
        ]
      }
    },
    "/analytics/reports/sales.csv": {
      "get": {
        "operationId": "AnalyticsController_exportSalesReportCsv",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "analytics"
        ]
      }
    },
    "/analytics/reports/inventory.csv": {
      "get": {
        "operationId": "AnalyticsController_exportInventoryReportCsv",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "analytics"
        ]
      }
    },
    "/analytics/overview": {
      "get": {
        "operationId": "AnalyticsController_getOverview",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get admin analytics overview"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "analytics"
        ]
      }
    },
    "/analytics/orders/status-breakdown": {
      "get": {
        "operationId": "AnalyticsController_getOrdersStatusBreakdown",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get order status breakdown analytics"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "analytics"
        ]
      }
    },
    "/analytics/products/top-selling": {
      "get": {
        "operationId": "AnalyticsController_getTopSellingProducts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get top-selling products for selected window"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "analytics"
        ]
      }
    },
    "/analytics/operations/fulfillment-sla": {
      "get": {
        "operationId": "AnalyticsController_getFulfillmentSla",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get fulfillment SLA performance metrics"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "analytics"
        ]
      }
    },
    "/analytics/payments/performance": {
      "get": {
        "operationId": "AnalyticsController_getPaymentsPerformance",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get payment performance metrics"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "analytics"
        ]
      }
    },
    "/analytics/promotions/efficiency": {
      "get": {
        "operationId": "AnalyticsController_getPromotionsEfficiency",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get promotions and coupon efficiency metrics"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "analytics"
        ]
      }
    },
...
    "/analytics/inventory/stockout-risk": {
      "get": {
        "operationId": "AnalyticsController_getStockoutRisk",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get stockout risk list based on sales velocity and current inventory"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "analytics"
        ]
      }
    },
    "/analytics/customers/retention": {
      "get": {
        "operationId": "AnalyticsController_getCustomersRetention",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get customer retention and repeat purchase metrics"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "analytics"
        ]
      }
    },
...
    "/analytics/funnel/source-attribution": {
      "get": {
        "operationId": "AnalyticsController_getSourceAttribution",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get source attribution metrics for storefront funnel"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "analytics"
        ]
      }
    },
    "/analytics/affiliates/performance": {
      "get": {
        "operationId": "AnalyticsController_getAffiliatePerformance",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get affiliate performance and commission KPIs"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "analytics"
        ]
      }
    },
    "/analytics/funnel/abandoned-carts": {
      "get": {
        "operationId": "AnalyticsController_getAbandonedCartsMetrics",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get abandoned cart and recovery KPI metrics"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "analytics"
        ]
      }
    },
    "/analytics/funnel/event-taxonomy": {
      "get": {
        "operationId": "AnalyticsController_getEventTaxonomy",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get tracked storefront event taxonomy and volume"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "analytics"
        ]
      }
    },
...
    "/analytics/quality/anomalies": {
      "get": {
        "operationId": "AnalyticsController_getAnomalies",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get analytics anomaly report versus previous window"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "analytics"
        ]
      }
    },
    "/analytics/quality/run-monitoring": {
      "post": {
        "operationId": "AnalyticsController_runMonitoring",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyticsGovernanceQueryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Run governance monitoring and emit critical anomaly alerts"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "analytics"
        ]
      }
    },
...
  },
  "info": {
    "title": "Alnjoom Telecom Store API",
    "description": "Backend APIs for Alnjoom Telecom Store",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateProductDto": {
        "type": "object",
        "properties": {}
      },
...
      "AnalyticsGovernanceQueryDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}